+++ tailfs/tailfsimpl/shared/pathutil.go
