+++ tailfs/tailfsimpl/shared/readonlydir.go
