+++ tailfs/tailfsimpl/shared/stat.go
