+++ tailfs/tailfsimpl/webdavfs/readonly_file.go
