+++ tailfs/tailfsimpl/webdavfs/stat_cache.go
