+++ tsd/tsd.go
@@ -38,17 +38,18 @@ import (
 
 // System contains all the subsystems of a Tailscale node (tailscaled, etc.)
 type System struct {
-	Dialer         SubSystem[*tsdial.Dialer]
-	DNSManager     SubSystem[*dns.Manager] // can get its *resolver.Resolver from DNSManager.Resolver
-	Engine         SubSystem[wgengine.Engine]
-	NetMon         SubSystem[*netmon.Monitor]
-	MagicSock      SubSystem[*magicsock.Conn]
-	NetstackRouter SubSystem[bool] // using Netstack at all (either entirely or at least for subnets)
-	Router         SubSystem[router.Router]
-	Tun            SubSystem[*tstun.Wrapper]
-	StateStore     SubSystem[ipn.StateStore]
-	Netstack       SubSystem[NetstackImpl] // actually a *netstack.Impl
-	TailfsForLocal SubSystem[*tailfs.FileSystemForLocal]
+	Dialer          SubSystem[*tsdial.Dialer]
+	DNSManager      SubSystem[*dns.Manager] // can get its *resolver.Resolver from DNSManager.Resolver
+	Engine          SubSystem[wgengine.Engine]
+	NetMon          SubSystem[*netmon.Monitor]
+	MagicSock       SubSystem[*magicsock.Conn]
+	NetstackRouter  SubSystem[bool] // using Netstack at all (either entirely or at least for subnets)
+	Router          SubSystem[router.Router]
+	Tun             SubSystem[*tstun.Wrapper]
+	StateStore      SubSystem[ipn.StateStore]
+	Netstack        SubSystem[NetstackImpl] // actually a *netstack.Impl
+	TailFSForLocal  SubSystem[tailfs.FileSystemForLocal]
+	TailFSForRemote SubSystem[tailfs.FileSystemForRemote]
 
 	// InitialConfig is initial server config, if any.
 	// It is nil if the node is not in declarative mode.
@@ -100,8 +101,10 @@ func (s *System) Set(v any) {
 		s.StateStore.Set(v)
 	case NetstackImpl:
 		s.Netstack.Set(v)
-	case *tailfs.FileSystemForLocal:
-		s.TailfsForLocal.Set(v)
+	case tailfs.FileSystemForLocal:
+		s.TailFSForLocal.Set(v)
+	case tailfs.FileSystemForRemote:
+		s.TailFSForRemote.Set(v)
 	default:
 		panic(fmt.Sprintf("unknown type %T", v))
 	}
