+++ tsnet/tsnet.go
@@ -530,7 +530,7 @@ func (s *Server) start() (reterr error) {
 	closePool.add(s.dialer)
 	sys.Set(eng)
 
-	// TODO(oxtoacart): do we need to support Tailfs on tsnet, and if so, how?
+	// TODO(oxtoacart): do we need to support TailFS on tsnet, and if so, how?
 	ns, err := netstack.Create(logf, sys.Tun.Get(), eng, sys.MagicSock.Get(), s.dialer, sys.DNSManager.Get(), sys.ProxyMapper(), nil)
 	if err != nil {
 		return fmt.Errorf("netstack.Create: %w", err)
