+++ tstest/integration/tailscaled_deps_test_linux.go
@@ -38,7 +38,7 @@ import (
 	_ "tailscale.com/ssh/tailssh"
 	_ "tailscale.com/syncs"
 	_ "tailscale.com/tailcfg"
-	_ "tailscale.com/tailfs"
+	_ "tailscale.com/tailfs/tailfsimpl"
 	_ "tailscale.com/tsd"
 	_ "tailscale.com/tsweb/varz"
 	_ "tailscale.com/types/flagtype"
