+++ tstest/integration/tailscaled_deps_test_windows.go
@@ -45,7 +45,7 @@ import (
 	_ "tailscale.com/safesocket"
 	_ "tailscale.com/syncs"
 	_ "tailscale.com/tailcfg"
-	_ "tailscale.com/tailfs"
+	_ "tailscale.com/tailfs/tailfsimpl"
 	_ "tailscale.com/tsd"
 	_ "tailscale.com/tsweb/varz"
 	_ "tailscale.com/types/flagtype"
